Version: 1.0.8
Added aria label to the social navigation.
Removed the title attribute from the Go to Top link.
Added translation functions for the fonts.
Moved the skip link to the wp_body_open filter.
Minor improvement for the font size setting.
Minor change to the position of the avatar in comments.
Removed function aeonblog_post_thumbnail which was oly used in one place.
Removed a reference to aeonblog-small-thumb because this custom image size was never set.